requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# OpenAI API
openai==1.3.7

# Environment and configuration
python-dotenv==1.0.0

# Vector database
chromadb==0.4.18

# Text processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Additional utilities
pydantic==2.5.0
typing-extensions==4.8.0
